+++ src/logic/buildings/SubsistenceFarm.cs
@@ -6,9 +6,9 @@ namespace EconSim.logic.buildings;
 
 public class SubsistenceFarm : Building
 {   
-    protected override int PRODUCTION_PER_PERSONYEAR { get; } = SimulationConstants.SubsistanceFarmValues.PRODUCTION_PER_PERSONYEAR;
+    protected override double PRODUCTION_PER_PERSONYEAR { get; } = SimulationConstants.SubsistanceFarmValues.PRODUCTION_PER_PERSONYEAR;
     protected override ItemType ITEM_CONSUMED { get; } = SimulationConstants.SubsistanceFarmValues.ITEM_CONSUMED;
-    protected override int ITEMS_CONSUMED_PER_UNIT_PRODUCED { get; } = SimulationConstants.SubsistanceFarmValues.CONSUMPTION_PER_UNIT_PRODUCED;
+    protected override double ITEMS_CONSUMED_PER_UNIT_PRODUCED { get; } = SimulationConstants.SubsistanceFarmValues.CONSUMPTION_PER_UNIT_PRODUCED;
     protected override CoinAmount WAGE_PER_PERSONYEAR { get; } = SimulationConstants.SubsistanceFarmValues.WAGE_PER_PERSONYEAR;
     protected override ItemType ITEM_PRODUCED { get; } = SimulationConstants.SubsistanceFarmValues.ITEM_PRODUCED;
